client/app/core/login/login.component.ts
import { Component, OnInit } from '@angular/core';
import { AuthService } from '../auth.service';

@Component({
  selector: 're-login',
  templateUrl: './login.component.html',
  styleUrls: ['./login.component.scss']
})
export class LoginComponent implements OnInit {

  auth0Redirect = true;
  auth0Remember = true;

  private readonly auth: AuthService;

  constructor(auth: AuthService) {
    this.auth = auth;
  }

  ngOnInit() {
    this.auth.displayLoginForm('login', this.auth0Redirect, this.auth0Remember);
  }

}
